Refactor login module packages
