+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/data/repository/login/LoginDataRepository.kt
@@ -1,6 +1,5 @@
-package com.gdavidpb.tuindice.login.data.login
+package com.gdavidpb.tuindice.login.data.repository.login
 
-import com.gdavidpb.tuindice.login.data.login.source.RemoteDataSource
 import com.gdavidpb.tuindice.login.domain.model.SignIn
 import com.gdavidpb.tuindice.login.domain.repository.LoginRepository
