+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/data/repository/login/SignInApi.kt
@@ -1,9 +1,9 @@
-package com.gdavidpb.tuindice.login.data.api
+package com.gdavidpb.tuindice.login.data.repository.login
 
 import com.gdavidpb.tuindice.base.domain.annotation.AttestedApi
 import com.gdavidpb.tuindice.base.domain.annotation.PublicApi
-import com.gdavidpb.tuindice.login.data.api.attestation.SignInAttestation
-import com.gdavidpb.tuindice.login.data.api.response.SignInResponse
+import com.gdavidpb.tuindice.login.data.repository.login.source.api.attestation.SignInAttestation
+import com.gdavidpb.tuindice.login.data.repository.login.source.api.response.SignInResponse
 import retrofit2.Response
 import retrofit2.http.Header
 import retrofit2.http.POST
